// front/src/app/models/livro.ts
export interface Livro {
    id: number;
    titulo: string;
    subtitulo: string;
    autor: string;
    editora: string;
    isbn: string;
    descricao: string;
    idioma: string;
    ano_publicacao: string;
    paginas: string;
    preco: string;
    estoque: string;
    desconto: string;
    disponivel: string;
    dimensoes: string;
    peso: string;
}